emotion-api/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart  # For image uploads
opencv-python-headless  # Optional, if using cv2